+++ src/main/kotlin/com/cmgapps/license/reporter/Report.kt
@@ -17,6 +17,8 @@
 package com.cmgapps.license.reporter
 
 import com.cmgapps.license.model.Library
+import groovy.lang.Closure
+import org.gradle.api.Action
 import org.gradle.api.Task
 import org.gradle.api.file.RegularFileProperty
 import org.gradle.api.provider.Property
@@ -24,6 +26,7 @@ import org.gradle.api.resources.TextResource
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.OutputFile
+import org.gradle.util.internal.ConfigureUtil
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KProperty
 import kotlin.reflect.javaType
@@ -32,7 +35,7 @@ abstract class Report(protected val libraries: List<Library>) {
     abstract fun generate(): String
 }
 
-open class LicensesReport(type: ReportType, task: Task) {
+open class LicensesReport(internal val type: ReportType, task: Task) {
     @get:Internal
     internal val name: String = type.name
 
@@ -48,11 +51,28 @@ open class LicensesReport(type: ReportType, task: Task) {
             task.project.buildDir.resolve("reports/licenses").resolve(task.name).resolve("licenses$extension")
         )
     }
+
+    internal open fun configure(
+        config: (Action<in LicensesReport>)? = null,
+        configHtml: (Action<in CustomizableHtmlReport>)? = null,
+        configCustom: (Action<in CustomizableReport>)? = null,
+    ) {
+        config?.execute(this)
+    }
 }
 
 class CustomizableHtmlReport(type: ReportType, task: Task) : LicensesReport(type, task) {
     @get:Input
     var stylesheet: Property<TextResource?> = task.project.objects.property(TextResource::class.java)
+
+    override fun configure(
+        config: (Action<in LicensesReport>)?,
+        configHtml: (Action<in CustomizableHtmlReport>)?,
+        configCustom: (Action<in CustomizableReport>)?,
+    ) {
+        super.configure(config, configHtml, configCustom)
+        configHtml?.execute(this)
+    }
 }
 
 class CustomizableReport(type: ReportType, task: Task) : LicensesReport(type, task) {
@@ -63,6 +83,15 @@ class CustomizableReport(type: ReportType, task: Task) : LicensesReport(type, ta
     fun generate(block: CustomReportAction? = null) {
         action = block
     }
+
+    override fun configure(
+        config: (Action<in LicensesReport>)?,
+        configHtml: (Action<in CustomizableHtmlReport>)?,
+        configCustom: (Action<in CustomizableReport>)?,
+    ) {
+        super.configure(config, configHtml, configCustom)
+        configCustom?.execute(this)
+    }
 }
 
 typealias CustomReportAction = (List<Library>) -> String
@@ -71,33 +100,123 @@ interface LicensesReportsContainer {
     @get:Internal
     val csv: LicensesReport
 
+    fun csv(config: Action<LicensesReport>)
+    fun csv(config: Closure<LicensesReport>)
+
     @get:Internal
     val html: CustomizableHtmlReport
+
+    fun html(config: Action<CustomizableHtmlReport>)
+    fun html(config: Closure<CustomizableHtmlReport>)
 
     @get:Internal
     val json: LicensesReport
 
+    fun json(config: Action<LicensesReport>)
+    fun json(config: Closure<LicensesReport>)
+
     @get:Internal
     val markdown: LicensesReport
+
+    fun markdown(config: Action<LicensesReport>)
+    fun markdown(config: Closure<LicensesReport>)
 
     @get:Internal
     val text: LicensesReport
 
+    fun text(config: Action<LicensesReport>)
+    fun text(config: Closure<LicensesReport>)
+
     @get:Internal
     val xml: LicensesReport
 
+    fun xml(config: Action<LicensesReport>)
+    fun xml(config: Closure<LicensesReport>)
+
     @get:Internal
     val custom: CustomizableReport
+
+    fun custom(config: Action<CustomizableReport>)
+    fun custom(config: Closure<CustomizableReport>)
+
+    @Internal
+    fun getAll(): List<LicensesReport>
 }
 
 internal class LicensesReportsContainerImpl(private val task: Task) : LicensesReportsContainer {
     override val csv: LicensesReport by LicenseReportDelegate(ReportType.CSV)
+    override fun csv(config: Action<LicensesReport>) {
+        csv.configure(config)
+    }
+
+    override fun csv(config: Closure<LicensesReport>) {
+        ConfigureUtil.configure(config, csv)
+    }
+
     override val html: CustomizableHtmlReport by LicenseReportDelegate(ReportType.HTML)
+
+    override fun html(config: Action<CustomizableHtmlReport>) {
+        html.configure(configHtml = config)
+    }
+
+    override fun html(config: Closure<CustomizableHtmlReport>) {
+        ConfigureUtil.configure(config, html)
+    }
+
     override val json: LicensesReport by LicenseReportDelegate(ReportType.JSON)
+    override fun json(config: Action<LicensesReport>) {
+        json.configure(config)
+    }
+
+    override fun json(config: Closure<LicensesReport>) {
+        ConfigureUtil.configure(config, json)
+    }
+
     override val markdown: LicensesReport by LicenseReportDelegate(ReportType.MARKDOWN)
+    override fun markdown(config: Action<LicensesReport>) {
+        markdown.configure(config)
+    }
+
+    override fun markdown(config: Closure<LicensesReport>) {
+        ConfigureUtil.configure(config, markdown)
+    }
+
     override val text: LicensesReport by LicenseReportDelegate(ReportType.TEXT)
+    override fun text(config: Action<LicensesReport>) {
+        text.configure(config)
+    }
+
+    override fun text(config: Closure<LicensesReport>) {
+        ConfigureUtil.configure(config, text)
+    }
+
     override val xml: LicensesReport by LicenseReportDelegate(ReportType.XML)
+    override fun xml(config: Action<LicensesReport>) {
+        xml.configure(config)
+    }
+
+    override fun xml(config: Closure<LicensesReport>) {
+        ConfigureUtil.configure(config, xml)
+    }
+
     override val custom: CustomizableReport by LicenseReportDelegate(ReportType.CUSTOM)
+    override fun custom(config: Action<CustomizableReport>) {
+        custom.configure(configCustom = config)
+    }
+
+    override fun custom(config: Closure<CustomizableReport>) {
+        ConfigureUtil.configure(config, custom)
+    }
+
+    override fun getAll(): List<LicensesReport> = listOf(
+        csv,
+        html,
+        json,
+        markdown,
+        text,
+        xml,
+        custom
+    )
 
     private inner class LicenseReportDelegate<T : LicensesReport>(private val reportType: ReportType) :
         ReadOnlyProperty<Any?, T> {